requirements.txt
rich>=13.7.0
python-nmap>=0.7.1
